api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
pandas==2.2.3
pydantic==2.9.2
python-dateutil==2.9.0.post0
